Small changes to networks
